Milestone 2: February 2025 Update
This commit renames and restructures a number of files, so things may be structured differently to before.

In more technical terms, most stuff has been moved out of dllmain and into more appropriately-named locations, which remain subject to further change.

Please review the README carefully to see known issues and bugs before filing any bug reports in the Discord server.

The main changes functionally are as follows:
- Windows 11 21H2 to 23H2 are now officially supported (+ the unsupported build warning will not display).
- Flyouts are handled better on non-immersive mode; volume will open the old Win7-era flyout while network opens Network and Sharing Center in Control Panel. The battery flyout will be accounted for in a future update.
- The option to disable taskbar pinning has been added. See README's registry options for details on how this works.
- Using hardware keys for brightness/volume will no longer crash explorer when running under immersive mode in Windows 11.
- Invoking context menus on the desktop or start menu for executable files (or their associated shortcuts) will no longer cause explorer to freeze on Windows 11.
- Handling for Win7-era msstyles has been rewritten and should be lighter on hardware resources and CPU.
- RPEnabled registry option has been removed - now, we automatically use the Win8-era msstyle classes if they are present in the provided style.
- Unnecessary internal check for thumbnail HWND has been removed which should allow desktop to load on *unsupported* uses of other explorer versions.
- Compatibility fixes added for newer Windows 11 23H2 updates.
- The "Program Manager" blockage on sign-out will no longer occur under immersive mode.
- Fixed Windows 10 start menu removal for 1507 and 1511 under immersive mode.
- Taskbar context menus will now render correctly on Windows 11.
- Win7-era msstyles will now load on Windows 8.1.

This is not an exhaustive list of the changes that have been made, but this is (at the very least) the vast majority of them.

A lot has gone into more appropriately reorganising the codebase and this is an ongoing effort.

This all being said, we hope that you will enjoy the latest Milestone 2 preview.

Thank you.