Milestone 2: February 2025 Update 2
This commit is focused mainly on clearing up legacy code that is no longer used or needed, but it does also address a couple of long-standing issues.

The following changes have been made:
- Taskbar selection behaviour under immersive mode has been fixed.
- The XamlOpenWith dialog will now appear on Windows 11 23H2+ as appropriate, when immersive mode is enabled.
- Checks for Nickel-based builds of Windows 11 are now applied to the base build, rather than to 23H2 specifically.

Please continue to report bugs to us, as and when you encounter them ahead of the full Milestone 2 release.

Thank you.