Milestone 2 ✅
This brings the publicly-available codebase in line with the final Milestone 2 release.

Additional details about licensing have been added to the README for clarity. This was already our policy as communicated on Discord and in software releases, but we also felt that it should be made clear here, given that it has always applied to uses of compiled source code. Nothing has changed.